Remove manual call to Logger#__destruct().
The destructor was called in EngineBlock's shutdown function to
let the Mail writer send its mails before it would die(). As there is
no apparent reason the shutdown function should die() and thus prevent
other shutdown functions to be run, the die() is hereby removed. This
allows other shutdown functions to be run, like the one in Monolog's
MailHandler. Before merging this commit upstream, this change must be
verified to be OK.