Configures logger using existing configuration
In order to leverage the monolog bundle for easy logging within SF,
existing (ini) configuration values are loaded as parameters and
used to configure the logger service.

In order to log to a preconfigured channel,
the default logger service is an realiased as the logger that
logs to that channel.