Removing PHPCPD due to low gains and conflicting vendors
After bumping min PHP version to 5.6 a newer, better version
of PHPunit can be used. However, this depends on a package
that depends on a package that PHPCPD also uses, and due to
the requirements of PHPCPD it cannot be updated. This means
that PHPunit cannot be updated. The actual value of PHPCPD
is fairly low, thus removig it in favor of an update of
PHPunit is an acceptable choice.
Alternatively, phpcpd could be installed globally on the
dev and test envs, or a phar could be used. Neither seems a
good solution for portability and maintainability.

If in the future PHPCPD depends on sebastion/verion:^2.0
then CPD could be reintroduced if that provides gains.