Pass $response by reference to the filter commands.
Some filter commands, notably RunAttributeManipulations, need to be able
to modify the response itself, not only the attributes. For example to
change the Subject NameId inside an attribute manipulation.

This change is needed to restore this functionality on PHP >= 5.4, as
this worked previously because all supplied variables were passed
by reference. From http://nl.php.net/call_user_func_array:
"Before PHP 5.4, referenced variables in param_arr are passed to the
function by reference, regardless of whether the function expects the
respective parameter to be passed by reference."