Split eduPersonPrincipalName only by first @ sign.
In the wild we sometimes encounter values like "user@voorbeeld.nl@example.org".
It's debatable whether the scope is then "voorbeeld.nl@example.org" or
just "example.org".

eduPerson 201310 recommends in section 2.2.10 on eduPersonScopedAffiliation to
split by the leftmost '@' sign in case of multuple @ signs. Let's follow that
here as well.

In any case it should never be "voorbeeld.nl" for this case, which was the
previous behaviour.