Set AllowCreate correctly for SAML2
The SAML2 version bump contains a change in the way AllowCreate is to be set in SAML2: https://github.com/simplesamlphp/saml2/commit/aecda70e72970c8d66af1d18d5b762a8df5731ed.

We were setting this as a string and this caused AllowCreate not to be set anymore, while there are some IdPs out there that require it.
This PR aims to fix that.