Include polyfills on all pages
Previously we only loaded polyfills on IE8 and lower versions. To
enable polyfills for ES5, ES6, HTML5 en DOM4 features we include the
polyfills in the javascript application file and let the libraries
decide if the polyfills are required.

The IE8 polyfill does not support loading in modern browsers, so to
keep things simpler we remove support for IE8.