Disallow authn requests signed with forbidden signature methods
EngineBlock now allows configuring one or more 'forbidden signature
methods'. If an authentication request is signed with such a method,
an error is logged and the generic error page is shown to the user.

Before this commit, the notice 'Received signed AuthnRequest from
Issuer...' was only logged when the signature was contained inside the
SAML message, but not when the signature was sent as a query
parameter. This commit changes this behaviour and logs the used
signature method regardless of how the signature was sent.