Do not process authn requests multiple times per request
The receiveRequest() and receiveResponse() methods on the bindings
module are costly methods that do logging, signature validation and
fetching of metadata. The methods do not cache their results so they
should never be called more than once per request.

Until commit 2ed6926, the receiveRequest() method was called just once
per request, by the Corto SingleSignOn module. Starting from that
commit, the method was called in two places, relying on caching by the
'internal bindings' mechanism to be efficient - which presumably has
never worked as intended for requests. In later commits, two more
invocations of the method were added causing signatures to be
validated four times for a single request.

This commit fixes the issue in three places:

1. The receiveRequest() method on the bindings module now stores the
   request object in the proxy server instance, making it available
   for other corto modules.

2. The SingleSignOn corto module uses the received request on the
   proxy server instead of calling receiveRequest() again.

3. The EngineBlock Corto adapter calls receiveRequest() just a single
   time, removing misleading comments about caching.