Update of saml-value-object library
The saml-value-object and beberlei assert libraries have been updated
to version 1.3.1 and 2.8.1 respectively.

The saml-value-object required version 2.6 and above of the assert lib,
this was introduced to make use of the isCallable method that was
introduced on the assert library in version 2.6. This allowed the
saml-value-object library to drop the home made implementation if favor
of beberlei's solution.

In doing so, engineblock-metadata was also bumped
one minor release to 2.5.0.