Make Profile url configurable in application.ini
A new configuration parameter was added to the application.ini. This
this config is read in the ProvideConsent service and injected into
the consent view. The profile url is then fed into the translation
helper who in turn sprintf's it into the translation string.

The nl and en translations have been updated.