Update to engineblock-metadata 3.1.0
This introduces two changes to EngineBlock:

1. Optional improved metadata caching

The configuration option "metadataRepository.database.type" can be set
to "CachedDoctrine", which behaves exactly like "Doctrine" but uses
per-request caching of query results. This strategy will likely become
the default behaviour in future versions.

2. Fix metadata cleanup after push

When pushing metadata to engineblock, sps or idps not present in the
pushed set of metadata were never deleted from the database. From now
on, any sp or idp not present in the pushed set of metadata is removed
after synchronization.