Import engineblock-metadata into engineblock
The engineblock-metadata repository version 3.1.0 is now part of the
EngineBlock repository. All classes previously namespaced under

    OpenConext\Component\EngineBlockMetadata

are now namespaced

    OpenConext\EngineBlock\Metadata

to conform to other engineblock components. No other changes have been
made. The goal of importing the repository into EngineBlock is to make
it easier to do minor refactoring of the code base.