Prevent scrolling to Consent button
When the consent button is out of view (smaller screens) the focus on
the button caused the screen to scroll down. Cutting of the top of the
page. This change tests if the provide consent button is visible on the
page, if it is. The focus is placed on it. Otherwise it isn't.