Rewrite behat tests to work without the janus pull client
The behat contexts (MockSpContext, MockIdpContext) use the
ServiceRegistryFixture object to write metadata to disk in the janus
V1 format. EngineBlock then reads this data re-using the janus pull
client. Since we don't want to support the janus pull mechanism, we
need to change how we provision metadata to EngineBlock in the tests.

Instead of writing an intermediary format to disk, we change the
ServiceRegistryFixture to create ServiceProvider and IdentityProvider
entities exactly as how they are used by EngineBlock. These entities
are serialized, saved to disk and loaded in an in-memory metadata
repository by EngineBlock.

The previous implementation allowed fine-grained access control in
scenario's. For example, entities could be setup to use a whitelist or
a blacklist, and granting or denying access could be done both
ways. This reflected the way janus allows access control but since we
don't want to duplicate this non-trivial logic this commit changes how
access control in behat tests work. Internally, EngineBlock does not
store this fine-grained information. It only knows what identity
providers are allowed for a given service provider: the 'allowed idp
entity ids'.

This commit removes all fine-grained access control from behat
contexts. As before, the default is to allow access between all
entities created in a scenario. Unlike before, it is only possible to
'disconnect' two specific entities:

    And SP "SP" is not connected to IdP "IDP"

Which would previously be written as:

    And SP "SP" uses a whitelists for access control

The changes to SpProxy.feature are a good example of how the new way
of working with access control simplifies the features.