Restore persistent ID / SP uuid tables
Commit 047bcd8 added database migration 'Version20180202131013' where
the saml_persistent_id and service_provider_uuid tables were
dropped. They are essential to EngineBlock (part of the
NameIdResolver).

This commit fixes the issue. If you've execute the migration in the
meanwhile, take care to downgrade the migration. The most important
goal of this commit is to prevent data loss on future deployments.