Remove LogChunkParser references from Contexts
The IdP and SP contexts contained references to the no longer existing
LogChunkParser. Only the SP Context actually had a reference to this
class but in a condition that was never used. This condition
(spMayRunInTransparentModeIfIndicatedIn) has been removed from the
context.

While at it, other unused imports have also been cleaned up.