Removed unused Repository injection
When finding entities publishable in EduGain a method was used that
accepted an optional Metadata repository (I think to be able to find
EG entities in yet another repository). The thing is, none of the
current MetadataRepositoryInterface implementations use this
possibility. And in addition this feature caused the invoke method on
the CachedDoctrineMetadataRepository to raise a fatal error (as the
MetadataRepositoryInterface cannot be serialized.

The solution to the problem was to update the signature for the
findEntitiesPublishableInEdugain method in the
MetadataRepositoryInterface and updating the current implementations.