Remove the composite metadata repository
Metadata always has one single source: in-memory (tests) or doctrine.

The special case where the workflowstate of EngineBlock is ignored in
all queries is solved in a different way now. Instead of cloning the
SP and IDP of EngineBlock, toggling the workflow state and then making
them available in the in-memory adapter, this commit removes all that
logic and simply changes the RemoveOtherWorkflowStatesFilter to take
into account the entity ID of EngineBlock (remove other workflow
states only if the SP/IDP is not EB).