Take scope into account when determining candidate IDPs
In case were there are only a few candidate IDPs based on scoping,
simply look up those entity IDs instead of fetching all entity IDs and
comparing them later.

The net effect of this change if probably not impressive for most use
cases, since the call to 'findAllIdentityProviderEntityIds' is already
filtered on the access configuration (allowed idps) of the service
provider.