Don't store SAML responses in user sessions
EngineBlock stores all SAML responses from the IDP in the user
session. Years ago, those responses were re-purposed on subsequent
logins. In order to support key roll-over EngineBlock also had to save
and restore the ID of the used signing key when re-using an old response.

Repurposing of responses is tricky in several ways so that feature was
removed (it's not EngineBlocks responsibility, the IDP should decide
if a session is still valid). But the SAML response is still stored
in the session, and the old signing key ID still overrides active
configuration.

This commit refactors the response cache mechanism:

 - the SAML response and ID of the used signing key is not saved to
   the session
 - leaving only the SP/IDP entity ID combination in the user
   session (so no more distinction between IN/OUT)
 - signing key used is always the key of the most recent/active configuration
 - rename methods to reflect that we are 'remembering the last used
   IDP' instead of 'sending a cached response'
 - after determining the 'remembered idp', resume the normal SSO flow
   but scoped to the remembered IDP instead of branching into a
   different code path to send the authentication request to the IDP