Replace the environment in app_dev.php for functional tests
Previously all behat tests were running using the dev environment. We
now need it to run in the test environment so we can have different
database configuration.

EngineBlock does not support running with 'app_test.php' in the URL,
and configuring a new vhost just for this purpose is more work than we
want to do now. So we simply replace the env in app_dev.php.