Remove the 'in-dev-machine' targets
Running ant tasks from host inside the guest does not work with the
new testing setup: create-test-db must too run inside the VM, and in
order to get that to work lots of ant tasks need to be duplicated.

Since we don't use the tasks anyway, this commit removes the
functionality and requires developers to run ant taks inside the
VM. So from now on, calling from inside the VM:

    ant behat-dev

Will enable functional testing, set the environmet in app_dev.php to
test, prepare the environment, run behat, and restore everything to
the development settings afterwards.

To restore the environment after cancelling the task:

    ant disable-functional-testing
    ant disable-test-env
    ant prepare-env