Simplify SP creation in Api test
When building a test SP with ARP. The test no longer calls the entire
constructor of the SP entity. Only the required entityId is passed to
the constructor. And the arp config is then set on the public arp field
of the SP entity.

This prevents us from updating the constructor call every time it
changes.