Remove the GNU make Makefile
While using make this way for running development-related tasks is a
nice idea, in practice we still use ant (build.xml). The Makefile is
outdated (was never updated with recent changes) and needs more work
to be usable again.

This commit removes the Makefile. If in the future we whish to
reintroduce the makefile again, it should be updated and replace
build.xml instead of having to build scripts to maintain.