Prevent unfiltered results in cached metadata repository
Cached results are only valid for a specific combination of filters
and visitors. If a filter or visitor is appended, the previously
cached results should be discarded. In practice, visitor/filter setup
is only done in the beginning of the request so resetting the cache
has little impact on the total number of queries.

While no bug has been found in EB5.7 (all queries are executed after
setting up the filters) not solving this issue in the repository will
open the door for future bugs if some metadata is queried before
setting up the filters.