Prove attribute manipulation with arp is broken
This test combines attribute manipulation and applying the arp in such
a way that causes the SAML2 library to throw the dreaded "Array of value
types and array of values have different size" error.

The next action will be to fix this in the SAML2 library.

(cherry picked from commit 7e5c956)