Show attribute information text in tooltip on consent screen
The attribute release policy (ARP) in the metadata of a service
provider contains a 'motivation' field explaining why the service
provider requires this attribute.

This commit addes a tooltip on the consent screen presenting the
motivation to the user.

Manage version 2.0.18 or higher is required for the motivations to
show up.

See: https://www.pivotaltracker.com/story/show/158146257