Implement a minimal variant of the consent screen
Manage before 2.0.18 sends a list of service provider entity IDs for
each identity provider indicating when not to ask for
consent. In newer versions of 2.0.18 manage sends configuration
objects for each service provider:

    "disable_consent_connections": [
    {
        "name": "https://serviceregistry.acc.surfconext.nl/simplesaml/module.php/saml/sp/metadata.php/default-sp",
        "type": "minimal_consent",
        "explanation:nl": "Explanation EN",
        "explanation:en": "Explanation NL"
    }

This commit changes the metadata storage backend to persist this
configuration and make it available to the ProviderConsent corto
module. A stripped version of the consent screen is shown when the
consent type is set to 'minimal_consent'. The differences between the
default and minimal consent screens are small, but important for legal
reasons.

See https://www.pivotaltracker.com/story/show/158145547