Prevent unexpected behaviour when using new manage on old EB
Before manage 2.0.18, consent should be disabled for all service
providers listed in 'disable_consent_connections'. Starting from the
new version of manage, the objects listed contain a type indicating
default, minimal or no consent.

In order to prevent unexpected behaviour when configuring default or
minimal consent in manage, we now check if the type is present and if
so, read the value to see if it's set to 'no_consent'.