Stop showing information feedback on PDP error page
The PDP error page should nog display the information feedback table,
but display the error report as we did before improving the error
reporting.

To do so, the error template now renders it's information feedback table
in a block. This block is emptied in the authorization-policy-violation
Twig template.

See request for changes by thijskh on:
https://github.com/OpenConext/OpenConext-engineblock/pull/539

(cherry picked from commit ab3bfaf)