Pre-fetch the NameID to show on consent
The NameID is not yet transformed to the definitive format
that will be outputted to the SP. This however is what we
want to show on the consent screen. To achieve this goal,
the format is applied before applying the SetNameId filter
(which will be applied in the output filter, we are stil
in the input filter context at this point).

This does not have any hard consequences for the rest of
the filters other than that extra database traffic is
introduced. This will be addressed in an upcomming commit.