Update the consent screen with FA icons
The warning triangle svg is removed from the project and removed by a
Font Awesome counterpart. The _consent.css.sass file was updated to use
the warning icon. The same was applied for the chevron up/down and
target blank icons.

To get everything to work nicely in IE11, some changes where required
to the HTML. Anchors are now wrapping their text in a span to be able
to correctly render the underlines on the anchors. This entailed
a couple of changes to the translations and in the consent.html.twig.

Browser tests have been performed and showed consistent results
throughout the range of most important browsers. (IE11, FF, Chrome,
Safari on ios, and Chrome and Firefox on Android)