Fix wayf unconnected idp button hover
The unconnected IdP rows should show the white button with text
"request access" to make sure it's clear what happens when you click it.
The 'press enter to select'- text should dissapear from the buttons.
Also the highlighting should be changed and the latest input device
should control the highlighting.