Show SP name on "No connection between institution and service" page
When a user wants to log into a service that the IdP does not allow,
and the user is presented with an error page "No connection between
institution and service", the entityID of the SP is listed. Now also
the SP name is shown in this message because that's more meaningful.