Improve the way FA SCSS files are loaded
The previous solution that was used consisted of copy'ing all of the
SCSS files to the themes folder and then loading them in the
application.scss. During review @pablothedude was right in telling me
this is not necessary, as the stylesheets can be directly loaded into
the application stylesheet.

The commit then removes the previously added scss files and adds the
three required files in the application.sass

The Gruntfile.js is updated and no longer copies the font_awesome_scss

See https://github.com/OpenConext/OpenConext-engineblock/pull/579#pullrequestreview-142729519