Prevent adding empty 'return' hidden input field
The return field is filled with the 'return' value that can optionally
be set by the ProcessedAssertionConsumer or AssertionConsumer. When
sending the SAML response using the POST binding, EB now tests if the
setValue method exists on the SamlMessage which is of the annotated
type in most cases. The method usually exists, so even tho the setter
for 'value' might have never been called, it was still added.

The fix consists of an additional test if the value is not empty.

A simple behat test was added that verifies the return form field is not
added when it should not be.

See https://www.pivotaltracker.com/story/show/159466469

(cherry picked from commit 4d8aa35300290db03789795619a0dde4e6402820)