Add database health query to config
The monitor_database_health_check_query ini config is used to set the
Symfony yml param, which is included in the monitor config that was
added to the services.yml file.

By default a sensible query is set in the application.ini, this can be
overridden if so desired by the end user.