Substitute Symf class cache loading for composer
Deprecated as of 3.3, will be removed in 4.0. PHP 7.0 no longer benefits
from this performance upgrade. This commit is to try to find out if this
hurts a PHP 5.6 noticeably.

To remedy this possible performance issue, the optimize autoloader
option for composer has been enabled. Following the suggestion from the
Symfony docs:

https://symfony.com/doc/3.3/components/class_loader/cache_class_loader.html