Make Acs location scheme validation more robust
The previous check would never have worked as it did not actually check
the scheme but the URI in its entirety.

In addition to fixing this issue, the logic was encapsulated in a
protected method.

Finally, the configured 'allowedAcsLocationSchemes' are now retrieved
from the config that is loaded on the Bindings module instead of using
the hardcoded schemes.