Prevent crash when OriginalIssuer is null
When OriginalIssuer can not be read from the response, EB used to crash.
This is prevented by reporting on this issue and skipping the additional
logging that might have been requested.

This is a workaround to an issue that is described in greater detail in:
https://www.pivotaltracker.com/story/show/161346289

The actual solution to this problem is yet to be found.

(cherry picked from commit 447d6a8ab8bb4d1ae71daa00bd433957a6ae2ca5)