Generate Persistent NameID value for consent
When the NameId is persistent, the ProvideConsent controller will
utilize the NameIdResolver to load the Persistent name id value for the
service/user combination.

If the nameId is transient, the name id will not be displayed on the
consent screen.

And finally if the format is unspecified, the collabPersonId is
displayed on consent, no further change required for that (as the
SetNameId filter is not yet applied).

Note that the NameId changes are applied on a local copy, only used on
the consent screen.