Allow selecting unconnected IdP's with keyboard
By positioning the unconnected idp picker in the section of the dom
where the keyboard navigation is active (idp-picker (previously called
connected-idp-picker)), we are now able to also select the unconnected
IdP's with the keyboard.

In order to open the modal associated with a mouse click on the
unconnected idp, the keyboard listener was updated. And on every ENTER
press, we test if the user entered on a connected or unconnected IdP.

When pressing the unconnected IdP, the IdpPicker (renamed from Connected
IdpPicker) will retrieve the title and entity id from the focused IdP
table row.

https://www.pivotaltracker.com/story/show/159595638