Provide a WAYF test endpoint
This endpoint is only available in the test and dev environments and
can be configured to simulate the different features of the WAYF.

A seeder is used to create dummy SAML2 entities. For now the number of
IdPs that are created can be configured with query parameters. Other
features of the IdP are hard coded (name, entitiyid and so forth).

The other available parameters are described in the js_testing.md readme.