Install Jest and provide a proof of concept
The POC is a meaningless unit test of the UnconnectedIdpPicker. But
proves jest works.

In order to get Jest to work with ES6/ecmascript 2015, the babelrc
config needed to be created, instructing jest-babel to expect es6 code
in the tests.

For now tests are run using `npm run jest` or `yarn test`. The next
task will be to be able to run the tests on Travis CI.