Display the no_attributes in the IdP table
Previously this text was displayed in a separate H2 heading. After
feedback from @arnoutt, we decided ot move the text back into the
IdP attribute table.

Added benefit is that we lose some complexity in the Twig template.

In order to test this solution, empty the attribute array found on line
https://github.com/OpenConext/OpenConext-engineblock/blob/defa3583661113e59272c844c13e036dd7c0793e/library/EngineBlock/Corto/Module/Service/ProvideConsent.php#L81

The discussion can be found here:
https://www.pivotaltracker.com/story/show/163524882/comments/199834274