Update feedback table ordering
The error code should be displayed on top. Fixed this by sorting the
array containing the feedback on key. These keys are later translated,
so they might not be ordered perfectly alphabetically in the end. But
the result is that the error code resides on top of the list.

Changed this after feedback in:
https://www.pivotaltracker.com/story/show/164076266/comments/200632691