Render 405 error pages more elegantly
The http 405 (request method not allowed) error handling has been made
more verbose for the users of engine block. A custom error screen was
added where the user is informed of the error situation.

On this screen we do not show the error table as we do during the
authentication process.

Step 1 of: https://www.pivotaltracker.com/story/show/164121663