Remove dead code
This portion can no longer be executed as the Symfony routing already
restricted the GET request on the SP ASC endpoint. The entire binding
not supported exception handling is not yet cleaned up, as this
exception can also be thrown when sending a SAMLResponse back to the
SP. This could hypothetically also result in that exception.