Optimize local selenium tests
A custom tag is added in behat to allow running the webdriver
tests separately from the Goutte tests. Also a seperate task is added
to ant to facilitate this change. Now we start selenium before running
the webdriver tests and stop the selenium service when we are done. This
is needed for OC-deploy.