Update focus when previous selection is hidden
The update focus was not called when the previous selection list was
hidden from view, because all items where removed. This only was called
when the done button was pressed. This update ensures that when the last
delete button is pressed, we also call updateFocus on the IdP list.
Effectively preparing the IdP list for interaction with mouse and
keyboard.