Only show feedback info if the block has content
Some errors explicitly empty the feedbackInfo block. In these cases, the
intro text would be displayed but the block is empty as we only test
for the existence of the feedbackInfo from the session (and that is
present). So an additional test on the block contents ensures we will
not render a confusing intro text.