Make uncaught error page reloadable
That was previously also possible, but resulted in a new unique request
id, timestamp combination upon reload. Making it very hard for service
desk to match the request id with the actual error message in logs.

This is resolved by redirecting to the feedback controller. On this
controller a new action handles the 'unknown error'.

This action is intentionally not accessible on the authentication/feedback
base route. This because the unknown error might not be authentication
related.

See: https://www.pivotaltracker.com/story/show/164076480