Add test to verify reloading of 500 error page
One of the validators has been overridden as a test specific service
(test stand-in). With the added feature that is looks for a specific
cookie. If this cookie exists, the validator will throw a runtime
exception. This exception is the handled by EB as a generic 500 error.

The test verifies if the request id remains intact between page reloads.