Move and merge two 5.11 migrations
These previously where migrations to be run in 5.10. They have been dropped from that release and moved into the 5.11 release. This required them to be moved to a new migration with a new timestamp in order to allow running the migration without problem.

Task 3 from https://www.pivotaltracker.com/story/show/166435147