Improve feedback information word wrapping
By using overflow-wrap with the anywhere option, we should be able to
display longer strings in the feedback blocks without losing any text to
an overflow hidden option.

Overflow wrap instructs the render engine to find a way to break the
string whereever it sees fitting. This might result in gramatically
incorrect word-breaks. But this would be virtualy impossible to do
correct in the context of an entity id. The ability to read the entire
string (which was not possible with the previous text-overflow: ellipsis
solution) is a requirement for the support team to be able to find
possible problems with the entity id's.

See task 7 from: https://www.pivotaltracker.com/story/show/166437858