Move the SURFboard to left on < large viewports
When the viewport is smaller than large (1024px), then the surfboard is
moved to a position of '-8% left' of the viewport boundry. This makes
sure the image is well readable when the viewport size is smaller.

The SVG itself (not tracked in this repo) is updated, the sign is moved
to the left of the surfboard.

See point 1 of https://www.pivotaltracker.com/story/show/166437858