Prevent vertical scrollbar
The margin bottom caused a scrollbar to show when the content still
fitted entirily on the page. This margin was added to create a
separation between the content block and the bottom of the page (this
would render a bleu section between the bottom of the page and the
bottom of the content block.

Solves point 2 of https://www.pivotaltracker.com/story/show/166437858