Improve PDP error screen styling
The HR is now always shown on the error page

The institution message is styled to be non-bold and italic. This makes
it far more distinctive.

Finally only the 'error_authorization_policy_violation_info' is rendered
in a h2 element. The error_authorization_policy_violation_desc is no
longer wrapped in a h2.