Fix the keys
The engineblock keys used for SFO seemed to be coming from a lot of
different places on different environments. When implementing sfo only
the ci keys were used. So now this is refactored to use the same keys
for (functional-) test to make things much more transparent.

The behat tests are currently ran as root but this is not necessary
anymore as it wil be chmodded to 440 and chowned as user:root to reflect
the production situation much better.