Rename 'sfo' to 'Stepup'
In engine block, with the SFO epic. We introduced step up authentication
in EB by using the SFO capabilities of StepUp Gateway. We however, in
code name every step up action: SFO. This is the working name of this
feature an was thus used, even tho it did not correctly express the
actual concept.

This is now changed to prevent possible confusion in the future.