Specify identifier and error code in feedback.yml
In order to make the FeedbackController less repetative, the things
that where specified in the controller (page identifier and the error
code) are now configured in the feedback.yml.

Some of the most generic controller actions have been made redundant and
are removed in this commit. Next effort will be put into removing the
Twig templates that can also be further abstracted.