Fix the eb stepup gw key
When implementing SFO initially the ci test keys were used. This
however seemed to work out in Travis but not for local development.
The Behat tests were executed as the local Vagrant user while the
private key could only be read as the engine_fpm_user (engine) due
to how Openconext-deploy is configured.