Prevent WAYF c-button from floating left
The button, displayed when the edit previously chosen selection editor is
activated. Would float to the left when the viewport size decreased to
<= medium. Resulting in a not so elegantly rendered list of previously
chosen IdP's.

As a solution. The button now remains floated to the right, but still
retains his own row. This to ensure the IdP labels are able to use the
maximum width of the row.

This could be achieved by removing the old code that would correctly
float the request access button to the left, which also adjusted the
(x) button. This block could be removed as the request access buttons
where later removed from view on smaller viewports. And the
incriminating code was not cleaned up at that point.

See: https://www.pivotaltracker.com/story/show/166694766