Remove idp metadata with sp-entity-id from home
The EB homepage advertised the possibility to open IdP metadata with the
certificate of a specified SP. This feature was not longer supported in
Engine. At least, I was not able to reproduce the behaviour described
for this feature.

There was not a lot to clean up, so only the homepage links have been
removed.

See: https://www.pivotaltracker.com/story/show/168249086