Show custom error when metadata not found
The metadata actions where results can be modified using an sp entity id
would result in a generic 500 error. This is not terribly user friendly.
And this PR aims to make the error more clear to the end user by
providing a custom error message. This message is partly built up out of
the exception message (which states what entity id was referenced) and a
custom error message (translatable).

See: https://www.pivotaltracker.com/story/show/168258776